+++ packages/admin-api-client-data-provider/test/index.spec.js
@@ -1,17 +1,4 @@
-import {
-  config,
-  about,
-  settings,
-  behaviors,
-  behavior,
-  behaviorsModel,
-  fixtures,
-  fixture,
-  fixturesModel,
-  alerts,
-  alert,
-  alertsModel,
-} from "../index";
+import { config, about, settings, alerts, alert, alertsModel } from "../index";
 
 describe("Exported methods", () => {
   it("should include config", () => {
@@ -24,30 +11,6 @@ describe("Exported methods", () => {
 
   it("should include settings", () => {
     expect(settings).toBeDefined();
-  });
-
-  it("should include behaviors", () => {
-    expect(behaviors).toBeDefined();
-  });
-
-  it("should include behavior", () => {
-    expect(behavior).toBeDefined();
-  });
-
-  it("should include behaviorsModel", () => {
-    expect(behaviorsModel).toBeDefined();
-  });
-
-  it("should include fixtures", () => {
-    expect(fixtures).toBeDefined();
-  });
-
-  it("should include fixture", () => {
-    expect(fixture).toBeDefined();
-  });
-
-  it("should include fixturesModel", () => {
-    expect(fixturesModel).toBeDefined();
   });
 
   it("should include alerts", () => {