+++ packages/admin-api-client/test/index.spec.js
@@ -15,11 +15,6 @@ import {
   readCustomRoutesVariants,
   useRouteVariant,
   restoreRoutesVariants,
-  // legacy
-  readBehaviors,
-  readBehavior,
-  readFixtures,
-  readFixture,
 } from "../index";
 
 describe("react-admin-client methods used with node", () => {
@@ -34,17 +29,12 @@ describe("react-admin-client methods used with node", () => {
     describe("when there are alerts", () => {
       it("should return alerts", async () => {
         const alerts = await readAlerts();
-        expect(alerts.length).toEqual(2);
-      });
-
-      it("should return alert about legacy usage", async () => {
-        const alerts = await readAlerts();
-        expect(alerts[0].message).toEqual(expect.stringContaining("Legacy mocks enabled"));
+        expect(alerts.length).toEqual(1);
       });
 
       it("should return alert about mock not defined", async () => {
         const alerts = await readAlerts();
-        expect(alerts[1].message).toEqual(
+        expect(alerts[0].message).toEqual(
           expect.stringContaining('Option "mock" was not defined')
         );
       });
@@ -58,12 +48,12 @@ describe("react-admin-client methods used with node", () => {
         });
         await wait(2000);
         const alerts = await readAlerts();
-        expect(alerts.length).toEqual(6);
+        expect(alerts.length).toEqual(5);
       });
 
       it("alert about files error should exist", async () => {
         const alerts = await readAlerts();
-        const alertId = alerts[5].id;
+        const alertId = alerts[4].id;
         const alert = await readAlert(alertId);
         expect(alert.id).toEqual(alertId);
         expect(alert.message).toEqual(expect.stringContaining("Error loading mocks"));
@@ -71,110 +61,25 @@ describe("react-admin-client methods used with node", () => {
     });
 
     describe("when alerts are removed", () => {
-      it("should return only one alert about legacy usage", async () => {
+      it("should return no alerts", async () => {
         await updateSettings({
           path: "mocks",
           mock: "base",
         });
         await wait(2000);
         const alerts = await readAlerts();
-        expect(alerts.length).toEqual(1);
+        expect(alerts.length).toEqual(0);
       });
-    });
-  });
-
-  describe("when reading fixtures", () => {
-    it("should return current fixtures collection", async () => {
-      const fixtures = await readFixtures();
-      expect(fixtures.length).toEqual(2);
-    });
-
-    it("first fixture model should exist", async () => {
-      const fixtures = await readFixtures();
-      const fixtureId = fixtures[0].id;
-      const fixture = await readFixture(fixtureId);
-      expect(fixture.id).toEqual(fixtureId);
-    });
-
-    it("second fixture model should exist", async () => {
-      const fixtures = await readFixtures();
-      const fixtureId = fixtures[1].id;
-      const fixture = await readFixture(fixtureId);
-      expect(fixture.id).toEqual(fixtureId);
-    });
-  });
-
-  describe("when reading behaviors", () => {
-    it("should return behaviors collection", async () => {
-      const behaviors = await readBehaviors();
-      expect(behaviors.length).toEqual(2);
-    });
-
-    it("first behavior model should exist", async () => {
-      const behaviors = await readBehaviors();
-      const behaviorName = behaviors[0].name;
-      const behavior = await readBehavior(behaviorName);
-      expect(behavior.name).toEqual(behaviorName);
-    });
-
-    it("second behavior model should exist", async () => {
-      const behaviors = await readBehaviors();
-      const behaviorName = behaviors[1].name;
-      const behavior = await readBehavior(behaviorName);
-      expect(behavior.name).toEqual(behaviorName);
-    });
-
-    it("fixture of behavior base should exist", async () => {
-      const behaviors = await readBehaviors();
-      const fixtureId = behaviors[0].fixtures[0];
-      const fixture = await readFixture(fixtureId);
-      expect(fixture.id).toEqual(fixtureId);
-    });
-
-    it("first fixture of behavior user2 should exist", async () => {
-      const behaviors = await readBehaviors();
-      const fixtureId = behaviors[1].fixtures[0];
-      const fixture = await readFixture(fixtureId);
-      expect(fixture.id).toEqual(fixtureId);
-    });
-  });
-
-  describe("when reading settings", () => {
-    it("should return current behavior", async () => {
-      const settings = await readSettings();
-      expect(settings.behavior).toEqual("base");
-    });
-
-    it("current behavior should exist", async () => {
-      const settings = await readSettings();
-      const currentBehavior = await readBehavior(settings.behavior);
-      expect(currentBehavior.name).toEqual(settings.behavior);
     });
   });
 
   describe("when updating settings", () => {
-    it("should update current behavior", async () => {
-      await updateSettings({
-        behavior: "user2",
-      });
-      const settings = await readSettings();
-      expect(settings.behavior).toEqual("user2");
-    });
-
     it("should update current delay", async () => {
       await updateSettings({
         delay: 1000,
       });
       const settings = await readSettings();
       expect(settings.delay).toEqual(1000);
-    });
-
-    it("should update current behavior", async () => {
-      await updateSettings({
-        behavior: "base",
-      });
-      const settings = await readSettings();
-      expect(settings.behavior).toEqual("base");
     });
   });