+++ packages/admin-api-paths/index.js
@@ -7,7 +7,3 @@ export const MOCKS = "/mocks";
 export const ROUTES = "/routes";
 export const ROUTES_VARIANTS = "/routes-variants";
 export const MOCK_CUSTOM_ROUTES_VARIANTS = "/mock-custom-routes-variants";
-
-export const LEGACY = "/legacy";
-export const BEHAVIORS = "/behaviors";
-export const FIXTURES = "/fixtures";