+++ packages/plugin-admin-api/jest.config.js
@@ -26,7 +26,6 @@ module.exports = {
 
   // The glob patterns Jest uses to detect test files
   testMatch: ["<rootDir>/test/**/*.spec.js"],
-  // testMatch: ["<rootDir>/test/**/legacy-behaviors-api.spec.js"],
 
   // The test environment that will be used for testing
   testEnvironment: "node",