+++ packages/plugin-admin-api/src/support/constants.js
@@ -10,6 +10,5 @@ Unless required by applicable law or agreed to in writing, software distributed 
 
 module.exports = {
   ADMIN_API_PATH_OPTION: "adminApiPath",
-  ADMIN_API_DEPRECATED_PATHS_OPTION: "adminApiDeprecatedPaths",
   PLUGIN_NAME: "plugin-admin-api",
 };