+++ packages/plugin-admin-api/CHANGELOG.md
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 ### Removed
 
+## [unreleased]
+
+### Removed
+ - feat: BREAKING CHANGE. Remove legacy APIs under /legacy path
+
 ## [2.2.1] - 2022-03-28
 
 ### Changed